Switch to using Alacritty over Terminator
Alacrity seems to be a better fit for how I use a terminal over
Terminator. It's aimed at being simpler and faster. Terminator supports
multiple tabs and windows, but I use tmux for that.

I also messed around with color to convince Emacs to use 24-bit color,
which seems to be working well. I believe the "XXX-direct" TERM values
are the right way to tell Emacs which colors to use.