Set TERM for ssh using alias
I'm setting TERM locally to alacritty-direct. This term value is
generally not supported on remote machines. The lack of support prevents
running tmux unless I set TERM when I call tmux on the remote
machine. Rather than having to do that all the time, it makes sense to
set TERM for all SSH connections to something that is widely
supported. I've never had an issue with term-256color, so I'm going with
that.