Make ZSH startup much faster
I profiled ZSH by adding this to the top of my .zshrc:

    zmodload zsh/zprof

And this to the bottom:

    zprof

It revealed that NVM was taking a ton of time on startup. Since I don't
use it, I removed it. I also updated oh-my-zsh, which seems to make it
load faster as well. Previously it spent a lot of time checking the
clipboard.